+++ src/main/java/com/factual/driver/Query.java
@@ -170,8 +170,20 @@ public class Query implements Filterable {
    * @param circle The circle within which to bound the results.
    * @return this Query.
    */
-  public Query within(Circle circle) {
-    queryParams.setParam(Constants.FILTER_GEO, circle);
+  public Query within(Shape shape) {
+    queryParams.setParam(Constants.FILTER_GEO, shape);
+    return this;
+  }
+  
+  /**
+   * Adds a filter to return results that are geographically near the
+   * specified Point.
+   * 
+   * @param point The point for which all results should be geographically near.
+   * @return this Query.
+   */
+  public Query near(Point point) {
+	queryParams.setParam(Constants.FILTER_GEO, point);
     return this;
   }